Endpoint to update, delete or detail a specific Employee ID
If the client makes a get request with the refered ID, the server returns the employee data as a json. If the client makes a put request with the specific ID, the server will update the employee data and persist the change in the database. If the client makes a delete request, it will erase the employee data and persist in the database.